backend/requirements.txt
fastapi
uvicorn[standard]

bcrypt
python-dotenv
pydantic-settings
pydantic
python-multipart
passlib
passlib[bcrypt]
SQLAlchemy
asyncpg
psycopg2-binary
fastapi-users[sqlalchemy]
